+++ src/talk2powersystemllm/app/server/main.py
@@ -26,7 +26,9 @@ from langchain_core.messages import (
 )
 from langchain_core.runnables import RunnableConfig
 from langgraph.checkpoint.redis import AsyncRedisSaver
+from langgraph.graph.state import CompiledStateGraph
 from rdflib import Namespace, Variable
+from redis.asyncio import Redis, RedisCluster
 from starlette.responses import JSONResponse, FileResponse
 from ttyg.tools import AutocompleteSearchTool
 
@@ -47,10 +49,11 @@ from .about import (
 from .auth import conditional_security
 from .config import settings
 from .healthchecks import (
-    GraphDBHealthchecker,
     HealthChecks,
+    GraphDBHealthchecker,
+    RedisHealthchecker,
+    LLMHealthchecker,
 )
-from .healthchecks.redis_healthcheck import RedisHealthchecker
 from .logging_config import LoggingConfig
 from .trouble import get_trouble_html
 from ..models import (
@@ -85,12 +88,6 @@ API_DESCRIPTION = ("Talk2PowerSystem Chat Bot Application provides functionality
 gtg_info: GoodToGoInfo = None
 # noinspection PyTypeChecker
 about_info: AboutInfo = None
-# noinspection PyTypeChecker
-agent_factory: Talk2PowerSystemAgentFactory = None
-# noinspection PyTypeChecker
-confidential_app: msal.ConfidentialClientApplication = None
-# noinspection PyTypeChecker
-COGNITE_SCOPES: list[str] = None
 ctx_request = ContextVar("request", default=None)
 LoggingConfig.config_logger(settings.logging_yaml_file)
 trouble_html = get_trouble_html()
@@ -99,50 +96,46 @@ trouble_html = get_trouble_html()
 @asynccontextmanager
 async def lifespan(_: FastAPI):
     logging.info("Starting the application")
-    global agent_factory
-    global confidential_app
-    global COGNITE_SCOPES
 
-    redis_password = settings.redis.password
-    redis_auth = ""
-    if redis_password:
-        redis_auth = f"{settings.redis.username}:{redis_password.get_secret_value()}@"
-    async with AsyncRedisSaver.from_conn_string(
-        redis_url=f"redis://{redis_auth}{settings.redis.host}:{settings.redis.port}",
-        connection_args={
-            'socket_connect_timeout': settings.redis.connect_timeout,
-            'socket_timeout': settings.redis.read_timeout,
-            'health_check_interval': settings.redis.healthcheck_interval,
-        },
-        ttl={
-            "default_ttl": settings.redis.ttl,
-            "refresh_on_read": settings.redis.ttl_refresh_on_read,
-        }
-    ) as memory_saver:
-        await memory_saver.asetup()
-        RedisHealthchecker(memory_saver._redis)
+    async with (
+        create_redis_client() as redis_client,
+        AsyncRedisSaver.from_conn_string(
+            redis_client=redis_client,
+            ttl={
+                "default_ttl": settings.redis.ttl,
+                "refresh_on_read": settings.redis.ttl_refresh_on_read,
+            }
+        ) as redis_saver
+    ):
+        await redis_saver.asetup()
+        RedisHealthchecker(redis_client)
+
+        app.state.callbacks = [LLMHealthchecker(redis_client)]
+
         agent_factory = Talk2PowerSystemAgentFactory(
             settings.agent_config,
-            checkpointer=memory_saver
+            checkpointer=redis_saver,
         )
+        app.state.agent_factory = agent_factory
+
+        GraphDBHealthchecker(agent_factory)
+
         if settings.security.enabled and agent_factory.settings.tools.cognite:
             if not agent_factory.settings.tools.cognite.client_secret:
                 raise ValueError(
                     "Cognite client secret must be provided in order to register the Cognite tools, "
                     "when the security is enabled."
                 )
-            confidential_app = msal.ConfidentialClientApplication(
+            app.state.confidential_app = msal.ConfidentialClientApplication(
                 settings.security.client_id,
                 authority=settings.security.authority,
                 client_credential=agent_factory.settings.tools.cognite.client_secret,
             )
-            COGNITE_SCOPES = [f"{agent_factory.settings.tools.cognite.base_url}/.default"]
-
-        GraphDBHealthchecker(agent_factory.graphdb_client, agent_factory.graphdb_repository_id)
+            app.state.cognite_scopes = [f"{agent_factory.settings.tools.cognite.base_url}/.default"]
 
         scheduler = AsyncIOScheduler()
-        scheduler.add_job(update_gtg_info, 'interval', seconds=settings.gtg_refresh_interval)
-        scheduler.add_job(update_about_info, 'interval', seconds=settings.about_refresh_interval)
+        scheduler.add_job(update_gtg_info, "interval", seconds=settings.gtg_refresh_interval)
+        scheduler.add_job(update_about_info, "interval", seconds=settings.about_refresh_interval)
 
         scheduler.start()
         app.state.scheduler = scheduler
@@ -157,6 +150,29 @@ async def lifespan(_: FastAPI):
         logging.info("Destroying the application")
         scheduler.shutdown()
         logging.info("Scheduler is stopped")
+
+
+def create_redis_client() -> Redis | RedisCluster:
+    redis_password = settings.redis.password
+    redis_auth = ""
+    if redis_password:
+        redis_auth = f"{settings.redis.username}:{redis_password.get_secret_value()}@"
+    redis_url = f"redis://{redis_auth}{settings.redis.host}:{settings.redis.port}"
+    connection_kwargs = {
+        "socket_connect_timeout": settings.redis.connect_timeout,
+        "socket_timeout": settings.redis.read_timeout,
+        "health_check_interval": settings.redis.healthcheck_interval,
+    }
+    if settings.redis.is_a_cluster:
+        return RedisCluster.from_url(
+            redis_url,
+            **connection_kwargs,
+        )
+    else:
+        return Redis.from_url(
+            redis_url,
+            **connection_kwargs,
+        )
 
 
 app = FastAPI(
@@ -285,6 +301,7 @@ async def update_about_info() -> None:
 
 
 def get_about_ontologies() -> list[AboutOntologyInfo]:
+    agent_factory = app.state.agent_factory
     query = files("talk2powersystemllm.app.server.queries").joinpath("about_ontologies_query.rq").read_text()
     query_results, _ = agent_factory.graphdb_client.eval_sparql_query(
         agent_factory.graphdb_repository_id, query, validation=False
@@ -301,6 +318,7 @@ def get_about_ontologies() -> list[AboutOntologyInfo]:
 
 
 def get_about_datasets() -> list[AboutDatasetInfo]:
+    agent_factory = app.state.agent_factory
     query = files("talk2powersystemllm.app.server.queries").joinpath("about_datasets_query.rq").read_text()
     query_results, _ = agent_factory.graphdb_client.eval_sparql_query(
         agent_factory.graphdb_repository_id, query, validation=False
@@ -316,6 +334,7 @@ def get_about_datasets() -> list[AboutDatasetInfo]:
 
 
 def get_about_graphdb() -> AboutGraphDBInfo:
+    agent_factory = app.state.agent_factory
     query = files("talk2powersystemllm.app.server.queries").joinpath("about_graphdb_query.rq").read_text()
     query_results, _ = agent_factory.graphdb_client.eval_sparql_query(
         agent_factory.graphdb_repository_id, query, validation=False
@@ -334,6 +353,7 @@ def get_about_graphdb() -> AboutGraphDBInfo:
 
 
 def get_about_agent() -> AboutAgentInfo:
+    agent_factory = app.state.agent_factory
     tools: dict[str, dict[str, Any]] = dict()
     tools["sparql_query"] = {"enabled": True}
     tools["display_graphics"] = {"enabled": True}
@@ -446,24 +466,6 @@ async def get_auth_config(
     )
 
 
-def exchange_obo_for_cognite(user_access_token: str) -> dict:
-    result = confidential_app.acquire_token_silent(COGNITE_SCOPES, account=None)
-    if result:
-        logging.debug("Cognite token acquired.")
-        return result["access_token"]
-
-    logging.debug("Acquiring token for Cognite using OBO.")
-    result = confidential_app.acquire_token_on_behalf_of(
-        user_assertion=user_access_token,
-        scopes=COGNITE_SCOPES,
-    )
-    if "access_token" not in result:
-        error_message = f"Failed to obtain OBO token for Cognite: {result}"
-        logging.error(error_message)
-        raise HTTPException(status_code=401, detail=error_message)
-    return result
-
-
 # noinspection PyUnusedLocal
 @app.get(
     "/rest/chat/diagrams/{filename}",
@@ -505,19 +507,6 @@ async def diagrams(
     return FileResponse(file_path, headers={"Cache-Control": "private, max-age=3600"})
 
 
-def build_diagram_image_url(request: Request, filename: str) -> str:
-    return str(
-        (settings.frontend_context_path if settings.frontend_context_path != "/" else "") + \
-        request.app.url_path_for("diagrams", filename=filename)
-    )
-
-
-def build_gdb_visual_graph_url(link: str) -> str:
-    return agent_factory.settings.graphdb.base_url + \
-        ("" if agent_factory.settings.graphdb.base_url.endswith("/") else "/") + \
-        link + "&embedded=true"
-
-
 # noinspection PyUnusedLocal
 @app.post(
     "/rest/chat/conversations",
@@ -557,16 +546,55 @@ async def conversations(
     authorization: Annotated[str | None, Header()] = None,
     claims=Depends(conditional_security),
 ) -> ChatResponse:
+    cognite_obo_token = await get_cognite_obo_token(authorization)
+    agent = app.state.agent_factory.get_agent(cognite_obo_token)
+    conversation_id = await get_or_create_conversation(chat_request, agent)
+    logging.info(f"Conversation {conversation_id}: Input \"{chat_request.question}\"")
+    user_datetime_ctx.set(x_user_datetime)
+
+    start = time.time()
+    try:
+        return await run_agent_on_input(request, agent, conversation_id, chat_request.question)
+    finally:
+        logging.info(
+            f"Conversation {conversation_id}: Elapsed time: {time.time() - start:.2f} seconds"
+        )
+
+
+async def get_cognite_obo_token(authorization: str | None) -> str | None:
     cognite_obo_token = None
+    agent_factory = app.state.agent_factory
     if (settings.security.enabled and agent_factory.settings.tools.cognite and
         agent_factory.settings.tools.cognite.client_secret):
         token_result = exchange_obo_for_cognite(authorization)
         cognite_obo_token = token_result["access_token"]
+    return cognite_obo_token
 
-    agent = agent_factory.get_agent(cognite_obo_token)
 
-    user_datetime_ctx.set(x_user_datetime)
+def exchange_obo_for_cognite(user_access_token: str) -> dict:
+    confidential_app = app.state.confidential_app
+    cognite_scopes = app.state.cognite_scopes
+    result = confidential_app.acquire_token_silent(cognite_scopes, account=None)
+    if result:
+        logging.debug("Cognite token acquired.")
+        return result["access_token"]
 
+    logging.debug("Acquiring token for Cognite using OBO.")
+    result = confidential_app.acquire_token_on_behalf_of(
+        user_assertion=user_access_token,
+        scopes=cognite_scopes,
+    )
+    if "access_token" not in result:
+        error_message = f"Failed to obtain OBO token for Cognite: {result}"
+        logging.error(error_message)
+        raise HTTPException(status_code=401, detail=error_message)
+    return result
+
+
+async def get_or_create_conversation(
+    chat_request: ChatRequest,
+    agent: CompiledStateGraph,
+) -> str:
     conversation_id = chat_request.conversation_id
     if conversation_id:
         checkpoint = await agent.checkpointer.aget({"configurable": {"thread_id": conversation_id}})
@@ -574,16 +602,24 @@ async def conversations(
             raise ConversationNotFound(f"Conversation with id \"{conversation_id}\" not found.")
     else:
         conversation_id = "thread_" + str(uuid.uuid4())
+    return conversation_id
 
-    logging.info(f"Conversation {conversation_id}: Input \"{chat_request.question}\"")
-    runnable_config = RunnableConfig(configurable={"thread_id": conversation_id})
-    input_ = {"messages": [{"role": "user", "content": chat_request.question}]}
 
-    start = time.time()
+async def run_agent_on_input(
+    request: Request,
+    agent: CompiledStateGraph,
+    conversation_id: str,
+    question: str
+) -> ChatResponse:
     messages: list[Message] = []
     graphics: list[Graphic] = []
     sum_input_tokens, sum_output_tokens, sum_total_tokens = 0, 0, 0
 
+    runnable_config = RunnableConfig(
+        configurable={"thread_id": conversation_id},
+        callbacks=app.state.callbacks,
+    )
+    input_ = {"messages": [{"role": "user", "content": question}]}
     # noinspection PyTypeChecker
     async for output in agent.astream(input_, runnable_config, stream_mode="updates"):
         output = dict(output)
@@ -623,12 +659,9 @@ async def conversations(
                         )
                     elif isinstance(tool_message.artifact, GraphDBVisualGraphArtifact):
                         graphics.append(
-                            VizGraphGraphic(type="vizGraph", url=build_gdb_visual_graph_url(tool_message.artifact.link))
+                            VizGraphGraphic(type="vizGraph",
+                                            url=build_gdb_visual_graph_url(tool_message.artifact.link))
                         )
-
-    logging.info(
-        f"Conversation {conversation_id}: Elapsed time: {time.time() - start:.2f} seconds"
-    )
 
     total_input_tokens = sum([message.usage.prompt_tokens for message in messages])
     total_output_tokens = sum([message.usage.completion_tokens for message in messages])
@@ -640,6 +673,20 @@ async def conversations(
         usage=Usage(completionTokens=total_output_tokens, promptTokens=total_input_tokens,
                     totalTokens=total_total_tokens)
     )
+
+
+def build_diagram_image_url(request: Request, filename: str) -> str:
+    return str(
+        (settings.frontend_context_path if settings.frontend_context_path != "/" else "") + \
+        request.app.url_path_for("diagrams", filename=filename)
+    )
+
+
+def build_gdb_visual_graph_url(link: str) -> str:
+    agent_factory = app.state.agent_factory
+    return agent_factory.settings.graphdb.base_url + \
+        ("" if agent_factory.settings.graphdb.base_url.endswith("/") else "/") + \
+        link + "&embedded=true"
 
 
 class ConversationNotFound(Exception):
@@ -715,7 +762,7 @@ async def explain(
 
 
 async def get_all_messages(conversation_id: str, message_id: str) -> list[Any]:
-    checkpoint = await agent_factory.checkpointer.aget({"configurable": {"thread_id": conversation_id}})
+    checkpoint = await app.state.agent_factory.checkpointer.aget({"configurable": {"thread_id": conversation_id}})
     if not checkpoint:
         raise ConversationNotFound(f"Conversation with id \"{conversation_id}\" not found.")
 
@@ -768,6 +815,7 @@ async def build_query_methods(
     executed_queries: dict[str, dict[str, str]],
     tools_calls_errors: dict[str, Any]
 ) -> list[QueryMethod]:
+    agent_factory = app.state.agent_factory
     query_methods: list[QueryMethod] = []
     for message in explain_messages:
         if isinstance(message, AIMessage) and message.tool_calls != []: