+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ============
 
 * [#349](https://github.com/statnett/Talk2PowerSystem_PM/issues/349):
-  Update `ttyg` to `4.1.0`, which uses the GraphDB client from rdflib 
+  Update `ttyg` to `4.2.0`, which uses the GraphDB client from rdflib 
   to solve the known issue that the SPARQL queries are executed 
   against GraphDB without using a connection pool.
   As a result the configuration `graphdb.sparql_timeout` is no longer available.
@@ -29,6 +29,14 @@
   Update the response from the `explain` endpoint to include additional fields 
   "advanced" and "hideArgs" to indicate the UI that a tool call or its 
   arguments by default shouldn't be displayed.
+* [#315](https://github.com/statnett/Talk2PowerSystem_PM/issues/315):
+  Implement additional checks for the GraphDB health:
+  - Call `/repositories/{repositoryId}/health` endpoint for the main GraphDB repository to catch long-running queries.
+  - Check the n-shot tool repository is present and healthy, if the n-shot tool is available.
+  - Check the n-shot tool connector is present and healthy, if the n-shot tool is available.
+* [#315](https://github.com/statnett/Talk2PowerSystem_PM/issues/315):
+  Implement LLM health check, which checks if any LLM errors were hit in the last 60 seconds.
+  The data is persisted in Redis.
 
 1.9.0-rc2
 ============